[css-contain] Size containment should ignore contents in flexbox/grid
The spec is pretty clear
(https://drafts.csswg.org/css-contain/#containment-size):
  "1. When laying out the containing element,
  it must be treated as having no contents."

The patch modifies ComputeIntrinsicLogicalWidths() and
FirstLineBoxBaseline() in LayoutFlexibleBox and LayoutGrid,
so the contents are ignored if the flex or grid container
has "contain: size".
Buttons are implemented as flexboxes internally,
so they get fixed too thanks to this change.

BUG=850169
TEST=external/wpt/css/css-contain/contain-size-button-001.html
TEST=external/wpt/css/css-contain/contain-size-flexbox-001.html
TEST=external/wpt/css/css-contain/contain-size-grid-001.html

Change-Id: Ia08c2c7b31126911cb4f0d1fc70aa444c22af8e3